# Add optimized quantize function for ARM